% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.arrays.1_to_500}
\alias{GPL200.arrays.1_to_500}
\alias{GPL200.arrays.501_to_1000}
\alias{GPL200.arrays.1001_to_1386}
\title{Data: A compendium of C. elegans Affymetrix microarry data}
\format{A numeric matrix with probesets in rows and arrays in columns.}
\usage{
GPL200.arrays.1_to_500

GPL200.arrays.501_to_1000

GPL200.arrays.1001_to_1386
}
\description{
A \code{22,625 x 1386} matrix containing Affymetrix microarray data gathered from the GEO
database for the model orgnaism \emph{C. elegans}.  The matrix was preprared by
running the \code{\link[affy]{rma}} algorithm on all of the arrays from a particular
experiment (arrays with the same GSE ID in GEO), followed by bias correction using
the \code{\link[bias]{biasCorrection}} algorithm.  Between-array quantile normalization
was not performed, as this tended to degrade the performance of the fastICA algorithm.
}
\details{
The original matrix has been split into three pieces in order to comply with
GitHub file size limits; users should combine the three matrices using \code{\link[base]{cbind}}
to work with the full matrix (see examples).
}
\examples{
\dontrun{
# Combine pieces into single matrix
x = cbind(GPL200.arrays.1_to_500, GPL200.arrays.501_to_1000,
  GPL200.arrays.1001_to_1386)
}

}
\keyword{datasets}
