% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.GO}
\alias{GPL200.GO}
\title{Data: A probeset:GO annotation matrix}
\format{A binary matrix with probeset IDs in rows and GO IDs in columns.}
\usage{
GPL200.GO
}
\description{
This is a binary probeset annotation matrix containing Affymetrix probeset IDs in rows and
Gene Ontology (GO) IDs in columns.  A value of 1 in the matrix indicates that the gene to
which the probeset has been mapped has been annotated with the corresponding GO term.  The
data was collected from the ENSEMBL database using the \code{biomaRt} package.
}
\keyword{datasets}
