% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.PromoterWords}
\alias{GPL200.PromoterWords}
\title{Data: A probeset:promoter word annotation matrix}
\format{A binary matrix with probeset IDs in rows and words in columns.}
\usage{
GPL200.PromoterWords
}
\description{
This is a binary probeset annotation matrix containing Affymetrix probeset IDs in rows and
predicted promoter regulatory words (oligonucleotides) in columns.  A value of 1 in the matrix
indicates that the gene to which the probeset has been mapped contains the corresponding word in
its 2-kb promoter region.  The promoter word dictionary was compiled using the MobyDick algorithm
(see \url{https://www.ncbi.nlm.nih.gov/pubmed/10977067}).
}
\keyword{datasets}
