% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.REACTOME}
\alias{GPL200.REACTOME}
\title{Data: A probeset:REACTOME annotation matrix}
\format{A binary matrix with probeset IDs in rows and REACTOME IDs in columns.}
\usage{
GPL200.REACTOME
}
\description{
This is a binary probeset annotation matrix containing Affymetrix probeset IDs in rows and
REACTOME IDs in columns.  A value of 1 in the matrix indicates that the gene to
which the probeset has been mapped has been annotated with the corresponding REACTOME pathway
or complex.  The data was collected from the ENSEMBL database using the \code{biomaRt} package.
}
\keyword{datasets}
