% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.S}
\alias{GPL200.S}
\title{Data: A set of predicted gene modules for the model organism C. elegans}
\format{A numeric matrix with probesets in rows and gene modules in columns.}
\usage{
GPL200.S
}
\description{
A \code{22,625 x 209} matrix containing loadings (weights) of each gene in 209
predicted gene modules.  Modules were predicted using the DEXICA package and
the GPL200.1386 compendium.
}
\keyword{datasets}
