% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.TissueExpr}
\alias{GPL200.TissueExpr}
\title{Data: A probeset:tissue annotation matrix}
\format{A binary matrix with probeset IDs in rows and tissues in columns.}
\usage{
GPL200.TissueExpr
}
\description{
This is a binary probeset annotation matrix containing Affymetrix probeset IDs in rows and
\emph{C. elegans} tissues in columns.  A value of 1 in the matrix indicates that the gene to
which the probeset has been mapped has been found to be expressed in the corresponding tissue.
The data was collected from the \href{http://gfpworm.org}{GFP Worm} database.
}
\keyword{datasets}
